introbook_backend/requirements.txt
Django==5.2.4
djangorestframework
django-cors-headers
psycopg2-binary
pandas
openpyxl